chore(ft_printf.h): remove unnecessary include of libft.h feat(ft_printf.h): add include of sys/types.h for ssize_t type feat(ft_printf.c): change len_args type from int to ssize_t for consistency with ft_printf_utils.c refactor(ft_printf.c): change function signature of ft_format and ft_puthexa to use ssize_t instead of int for len_args refactor(ft_printf_utils.c): change function signature of ft_printchar, ft_printstr, ft_printptr, ft_printdec, and ft_printudec to use ssize_t instead of int for len_args

+++ src/ft_printf.c
@@ -6,13 +6,13 @@
 /*   By: ladloff <ladloff@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/01 20:54:11 by ladloff           #+#    #+#             */
-/*   Updated: 2023/05/21 23:06:18 by ladloff          ###   ########.fr       */
+/*   Updated: 2023/05/21 23:40:00 by ladloff          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_format(va_list *ap, const char format, int *len_args)
+void	ft_format(va_list *ap, const char format, ssize_t *len_args)
 {
 	if (format == 'c')
 		ft_printchar(va_arg(*ap, unsigned int), len_args);
@@ -32,7 +32,7 @@ void	ft_format(va_list *ap, const char format, int *len_args)
 		ft_printchar('%', len_args);
 }
 
-void	ft_puthexa(unsigned long n, const char *base, int *len_args)
+void	ft_puthexa(unsigned long n, const char *base, ssize_t *len_args)
 {
 	int		i;
 	char	hex[16];
@@ -52,7 +52,7 @@ int	ft_printf(const char *format, ...)
 {
 	int		i;
 	va_list	ap;
-	int		len_args;
+	ssize_t	len_args;
 
 	i = -1;
 	len_args = 0;
@@ -64,5 +64,5 @@ int	ft_printf(const char *format, ...)
 		else
 			ft_printchar(format[i], &len_args);
 	}
-	return (len_args);
+	return ((int)len_args);
 }
